requirements.txt
aniso8601==2.0.1
appdirs==1.4.3
certifi==2018.1.18
chardet==3.0.4
click==6.7
decorator==4.2.1
deprecation==2.0
dogpile.cache==0.6.4
Flask==0.12.2
Flask-RESTful==0.3.6
futures==3.2.0
gevent==1.3a1
greenlet==0.4.13
idna==2.6
ipaddress==1.0.19
iso8601==0.1.12
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsonpatch==1.21
jsonpointer==2.0
keystoneauth1==3.4.0
MarkupSafe==1.0
munch==2.2.0
netifaces==0.10.6
openstacksdk==0.11.3
os-service-types==1.2.0
packaging==17.0
pbr==3.1.1
pyparsing==2.2.0
pytz==2018.3
PyYAML==3.12
requests==2.18.4
requestsexceptions==1.4.0
six==1.11.0
stevedore==1.28.0
urllib3==1.22
Werkzeug==0.14.1
